Add files via upload
This project allows me to learn how to build an upgradeable smart contract. Upgradability in Smart contract allows us to fix the contract when there are bugs. So users can still access to the functionality of the contract without deploying a new contract.

In this project, we pretend Dog.sol contains bugs. And DogsUpdated.sol is used to replace Dog.sol. Also, Storage.sol, which is the storage contract, contains mappings instead of variables